Update with info on how to pull from main branch
This is important if we want to make sure everyone has the most up to date code. If you skip this step and just do "git pull" you will only ever pull from your own forked repository and not get the latest updates that other members of the team may have made.